Refactor webgpu crate
split lib.rs into multiple modules
