+++ components/webgpu/identity.rs
@@ -2,51 +2,44 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use msg::constellation_msg::PipelineId;
+use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
 use serde::{Deserialize, Serialize};
 
 use crate::wgc::id::{
-    AdapterId, BindGroupId, BindGroupLayoutId, BufferId, CommandBufferId, ComputePipelineId,
-    DeviceId, PipelineLayoutId, QuerySetId, RenderBundleId, RenderPipelineId, SamplerId,
-    ShaderModuleId, StagingBufferId, SurfaceId, TextureId, TextureViewId,
+    AdapterId, BindGroupId, BindGroupLayoutId, BufferId, CommandBufferId, CommandEncoderId,
+    ComputePipelineId, DeviceId, PipelineLayoutId, QueueId, RenderBundleId, RenderPipelineId,
+    SamplerId, ShaderModuleId, SurfaceId, TextureId, TextureViewId,
 };
-use crate::{ErrorScopeId, WebGPUDevice};
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub enum WebGPUOpResult {
-    ValidationError(String),
-    OutOfMemoryError,
-    Success,
+macro_rules! webgpu_resource {
+    ($name:ident, $id:ty) => {
+        #[derive(Clone, Copy, Debug, Deserialize, Hash, PartialEq, Serialize)]
+        pub struct $name(pub $id);
+
+        impl MallocSizeOf for $name {
+            fn size_of(&self, _ops: &mut MallocSizeOfOps) -> usize {
+                0
+            }
+        }
+
+        impl Eq for $name {}
+    };
 }
 
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub enum WebGPUMsg {
-    FreeAdapter(AdapterId),
-    FreeDevice(DeviceId),
-    FreeBuffer(BufferId),
-    FreePipelineLayout(PipelineLayoutId),
-    FreeComputePipeline(ComputePipelineId),
-    FreeRenderPipeline(RenderPipelineId),
-    FreeBindGroup(BindGroupId),
-    FreeBindGroupLayout(BindGroupLayoutId),
-    FreeCommandBuffer(CommandBufferId),
-    FreeTexture(TextureId),
-    FreeTextureView(TextureViewId),
-    FreeSampler(SamplerId),
-    FreeSurface(SurfaceId),
-    FreeShaderModule(ShaderModuleId),
-    FreeRenderBundle(RenderBundleId),
-    FreeStagingBuffer(StagingBufferId),
-    FreeQuerySet(QuerySetId),
-    WebGPUOpResult {
-        device: WebGPUDevice,
-        scope_id: Option<ErrorScopeId>,
-        pipeline_id: PipelineId,
-        result: WebGPUOpResult,
-    },
-    CleanDevice {
-        device: WebGPUDevice,
-        pipeline_id: PipelineId,
-    },
-    Exit,
-}
+webgpu_resource!(WebGPUAdapter, AdapterId);
+webgpu_resource!(WebGPUBindGroup, BindGroupId);
+webgpu_resource!(WebGPUBindGroupLayout, BindGroupLayoutId);
+webgpu_resource!(WebGPUBuffer, BufferId);
+webgpu_resource!(WebGPUCommandBuffer, CommandBufferId);
+webgpu_resource!(WebGPUCommandEncoder, CommandEncoderId);
+webgpu_resource!(WebGPUComputePipeline, ComputePipelineId);
+webgpu_resource!(WebGPUDevice, DeviceId);
+webgpu_resource!(WebGPUPipelineLayout, PipelineLayoutId);
+webgpu_resource!(WebGPUQueue, QueueId);
+webgpu_resource!(WebGPURenderBundle, RenderBundleId);
+webgpu_resource!(WebGPURenderPipeline, RenderPipelineId);
+webgpu_resource!(WebGPUSampler, SamplerId);
+webgpu_resource!(WebGPUShaderModule, ShaderModuleId);
+webgpu_resource!(WebGPUSurface, SurfaceId);
+webgpu_resource!(WebGPUTexture, TextureId);
+webgpu_resource!(WebGPUTextureView, TextureViewId);
